Make `sendPasswordResetEmail` response null (#913)
## Description

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.



#### PR Dependency Tree


* **PR #913** 👈

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)
