+++ src/user-management/interfaces/send-password-reset-email-options.interface.ts
@@ -1,0 +1,9 @@
+export interface SendPasswordResetEmailOptions {
+  email: string;
+  passwordResetUrl: string;
+}
+
+export interface SerializedSendPasswordResetEmailOptions {
+  email: string;
+  password_reset_url: string;
+}
