+++ src/user-management/serializers/send-password-reset-email.serializer.ts
@@ -1,17 +1,7 @@
 import {
   SendPasswordResetEmailOptions,
-  SendPasswordResetEmailResponse,
-  SendPasswordResetEmailResponseResponse,
   SerializedSendPasswordResetEmailOptions,
 } from '../interfaces';
-import { deserializeUser } from './user.serializer';
-
-export const deserializeSendPasswordResetEmailResponse = (
-  sendPasswordResetEmailResponse: SendPasswordResetEmailResponseResponse,
-): SendPasswordResetEmailResponse => ({
-  token: sendPasswordResetEmailResponse.token,
-  user: deserializeUser(sendPasswordResetEmailResponse.user),
-});
 
 export const serializeSendPasswordResetEmailOptions = (
   options: SendPasswordResetEmailOptions,
