+++ src/user-management/user-management.spec.ts
@@ -345,12 +345,9 @@ describe('UserManagement', () => {
     });
   });
 
-  describe('/user_management/password_reset/send', () => {
+  describe('sendPasswordResetEmail', () => {
     it('sends a Send Password Reset Email request', async () => {
-      mock.onPost(`/user_management/password_reset/send`).reply(200, {
-        token: 'password-reset-token',
-        user: userFixture,
-      });
+      mock.onPost(`/user_management/password_reset/send`).reply(200);
       const resp = await workos.userManagement.sendPasswordResetEmail({
         email: 'test01@example.com',
         passwordResetUrl: 'https://example.com/forgot-password',
@@ -360,12 +357,7 @@ describe('UserManagement', () => {
         `/user_management/password_reset/send`,
       );
 
-      expect(resp).toMatchObject({
-        token: 'password-reset-token',
-        user: {
-          email: 'test01@example.com',
-        },
-      });
+      expect(resp).toBeUndefined();
     });
   });
